(snippets) Add snippet create, list, and partial support for delete
